refactor: Rule.class 패키지 위치 수정
