+++ src/main/java/chess/domain/Rule.java
@@ -1,4 +1,4 @@
-package chess.domain.RuleImpl;
+package chess.domain;
 
 import chess.domain.Position;
